# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/attendance": {
      "get": {
        "operationId": "AttendanceController_findAll",
        "parameters": [
          {
            "name": "sessionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance[] — sessionId·studentId·status"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "출결 목록(Attendance[]) — sessionId 지정 시 해당 세션만",
        "tags": [
          "attendance"
        ]
      },
      "put": {
        "operationId": "AttendanceController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "upsert된 Attendance"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "출결 기록(upsert) — (세션,학생) 1행. FK·유니크 무결성 보장",
        "tags": [
          "attendance"
        ]
      }
    },
...
    "/api/counsel": {
      "get": {
        "operationId": "CounselController_findForms",
        "parameters": [],
        "responses": {
          "200": {
            "description": "CounselForm[] — 신청자·상태·관심 과목/코스·다음 상담일 등"
          },
          "403": {
            "description": "강사 접근 불가 — manager/admin/super_admin 전용"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 접수 목록(CounselForm[])",
        "tags": [
          "counsel"
        ]
      },
      "post": {
        "operationId": "CounselController_createForm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCounselDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 CounselForm"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 접수 생성 [관리 역할] — 전체 폼 저장, status=requested, nextContactAt은 예약 캘린더 단일 소스",
        "tags": [
          "counsel"
        ]
      }
    },
    "/api/counsel/rounds": {
      "get": {
        "operationId": "CounselController_findRounds",
        "parameters": [
          {
            "name": "counselFormId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CounselRound[] — 회차·요약·결과·다음 액션"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 회차 목록(CounselRound[]). counselFormId로 필터. [관리 역할]",
        "tags": [
          "counsel"
        ]
      }
    },
    "/api/counsel/analytics/funnel": {
      "get": {
        "operationId": "CounselController_funnel",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "접수일 시작(YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "접수일 끝(YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 퍼널 집계 — 상태·회차 도달·이탈 회차·전환율·평균 소요(순수 함수 파생, 기간 필터) [관리 역할]",
        "tags": [
          "counsel"
        ]
      }
    },
    "/api/counsel/analytics/correlation": {
      "get": {
        "operationId": "CounselController_correlation",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "접수일 시작(YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "접수일 끝(YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담↔수강 과목 상관관계 — 희망 과목(관심 SSOT)×실제 등록 과목(enrollments) 조인 매트릭스·전환율 [관리 역할]",
        "tags": [
          "counsel"
        ]
      }
    },
    "/api/counsel/{id}": {
      "get": {
        "operationId": "CounselController_findForm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CounselForm"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 폼 단건(CounselForm) — 관리 역할 상세 화면용. 없는 id=404.",
        "tags": [
          "counsel"
        ]
      },
      "patch": {
        "operationId": "CounselController_updateForm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCounselDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 CounselForm"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 폼 수정 [관리 역할] — 전체 입력 및 nextContactAt 예약 캘린더 동기화",
        "tags": [
          "counsel"
        ]
      },
      "delete": {
        "operationId": "CounselController_removeForm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 폼과 회차 soft delete [관리 역할] — 감사 스냅샷 포함",
        "tags": [
          "counsel"
        ]
      }
    },
    "/api/counsel/{id}/aggregate": {
      "get": {
        "operationId": "CounselController_findAggregate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 폼+회차+연결 학생 aggregate [관리 역할]",
        "tags": [
          "counsel"
        ]
      }
    },
    "/api/counsel/{id}/rounds": {
      "post": {
        "operationId": "CounselController_createRound",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCounselRoundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 CounselRound"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 회차 추가 [관리 역할] — roundNo 자동, 폼 nextContactAt 동기화",
        "tags": [
          "counsel"
        ]
      }
    },
    "/api/counsel/{id}/rounds/{roundId}": {
      "patch": {
        "operationId": "CounselController_updateRound",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "roundId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCounselRoundDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 회차 수정 [관리 역할] — snapshot/다음 상담일 동기화+감사",
        "tags": [
          "counsel"
        ]
      },
      "delete": {
        "operationId": "CounselController_removeRound",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "roundId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "상담 회차 soft delete [관리 역할] — 최신 다음 상담일 재계산+감사",
        "tags": [
          "counsel"
        ]
      }
    },
...
  },
  "info": {
    "title": "TACO ERP API",
    "description": "TnAcademy 백오피스 운영 API. Postgres 영속 저장소와 역할 기반 접근 제어를 사용하는 실제 서비스 계약.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "현재 호스트"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "access_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "UpsertAttendanceDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "number",
            "minimum": 1
          },
          "studentId": {
            "type": "number",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "present",
              "late",
              "absent",
              "excused"
            ]
          }
        },
        "required": [
          "sessionId",
          "studentId",
          "status"
        ]
      },
...
      "CreateCounselDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "number"
          },
          "source": {
            "type": "string",
            "enum": [
              "internal_form",
              "naver_form",
              "google_form",
              "manual",
              "etc"
            ]
          },
          "submitterType": {
            "type": "string",
            "enum": [
              "parent",
              "student",
              "staff",
              "unknown"
            ]
          },
          "assignedStaffId": {
            "type": "number"
          },
          "referenceNotes": {
            "type": "string",
            "maxLength": 2000
          },
          "nextContactAt": {
            "type": "string"
          }
        },
        "required": [
          "studentId",
          "source"
        ]
      },
      "UpdateCounselDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "requested",
              "pending",
              "registered",
              "dropped"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "internal_form",
              "naver_form",
              "google_form",
              "manual",
              "etc"
            ]
          },
          "submitterType": {
            "type": "string",
            "enum": [
              "parent",
              "student",
              "staff",
              "unknown"
            ]
          },
          "studentId": {
            "type": "number"
          },
          "assignedStaffId": {
            "type": "number",
            "nullable": true
          },
          "referenceNotes": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          },
          "nextContactAt": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CounselFormSnapshotDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "number"
          },
          "assignedStaffId": {
            "type": "number",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "requested",
              "pending",
              "registered",
              "dropped"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "internal_form",
              "naver_form",
              "google_form",
              "manual",
              "etc"
            ]
          },
          "submitterType": {
            "type": "string",
            "enum": [
              "parent",
              "student",
              "staff",
              "unknown"
            ]
          },
          "referenceNotes": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          },
          "nextContactAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "studentId",
          "status",
          "source",
          "submitterType"
        ]
      },
      "CreateCounselRoundDto": {
        "type": "object",
        "properties": {
          "counselorId": {
            "type": "number"
          },
          "summary": {
            "type": "string",
            "maxLength": 300
          },
          "detail": {
            "type": "string",
            "maxLength": 2000
          },
          "result": {
            "type": "string",
            "enum": [
              "positive",
              "neutral",
              "negative",
              "no_response",
              "registered"
            ]
          },
          "nextAction": {
            "type": "string",
            "maxLength": 500
          },
          "nextContactAt": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "formSnapshot": {
            "$ref": "#/components/schemas/CounselFormSnapshotDto"
          }
        }
      },
      "UpdateCounselRoundDto": {
        "type": "object",
        "properties": {
          "counselorId": {
            "type": "number",
            "nullable": true
          },
          "scheduledAt": {
            "type": "string",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "nullable": true
          },
          "isCompleted": {
            "type": "boolean"
          },
          "summary": {
            "type": "string",
            "nullable": true,
            "maxLength": 300
          },
          "detail": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          },
          "result": {
            "type": "string",
            "nullable": true,
            "enum": [
              "positive",
              "neutral",
              "negative",
              "no_response",
              "registered"
            ]
          },
          "nextAction": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "nextContactAt": {
            "type": "string",
            "nullable": true
          },
          "formSnapshot": {
            "$ref": "#/components/schemas/CounselFormSnapshotDto"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "access_token": []
    }
  ]
}